# nestjs-dredd/openapi.yml
openapi: 3.0.0
paths:
  /:
    get:
      operationId: AppController_getHello
      parameters: []
      responses:
        '200':
          description: ''
  /members:
    post:
      operationId: MembersController_create
      summary: 単体登録API
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateMemberDto'
      responses:
        '201':
          description: 登録したメンバー設定を返却
          content:
            application/json; charset=utf-8:
              schema:
                $ref: '#/components/schemas/Member'
      tags: &ref_0
        - /members
    get:
      operationId: MembersController_findAll
      summary: 全体取得API
      parameters: []
      responses:
        '200':
          description: 登録済メンバー設定を複数返却
          content:
            application/json; charset=utf-8:
              schema:
                $ref: '#/components/schemas/Member'
      tags: *ref_0
  /members/{id}:
    get:
      operationId: MembersController_findOne
      summary: 単体取得API
      parameters:
        - name: id
          required: true
          in: path
          example: 1
          schema:
            type: number
      responses:
        '200':
          description: 指定されたIDのメンバー設定を返却
          content:
            application/json; charset=utf-8:
              schema:
                $ref: '#/components/schemas/Member'
      tags: *ref_0
    patch:
      operationId: MembersController_update
      summary: 単体更新API
      parameters:
        - name: id
          required: true
          in: path
          example: 1
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateMemberDto'
      responses:
        '200':
          description: 更新後のメンバー設定を返却
          content:
            application/json; charset=utf-8:
              schema:
                $ref: '#/components/schemas/Member'
      tags: *ref_0
    delete:
      operationId: MembersController_remove
      summary: 単体削除API
      parameters:
        - name: id
          required: true
          in: path
          example: 1
          schema:
            type: number
      responses:
        '200':
          description: 削除されたメンバーの設定を返却
          content:
            application/json; charset=utf-8:
              schema:
                $ref: '#/components/schemas/Member'
      tags: *ref_0
info:
  title: Members API docs
  description: MembersのAPI仕様書です。
  version: '1.0'
  contact: {}
tags: []
servers: []
components:
  schemas:
    CreateMemberDto:
      type: object
      properties: {}
    Member:
      type: object
      properties:
        id:
          type: number
          example: 1
          description: メンバーID
        name:
          type: string
          example: アルファ太郎
          description: メンバーの氏名
        age:
          type: number
          example: 25
          description: メンバーの年齢
      required:
        - id
        - name
        - age
    UpdateMemberDto:
      type: object
      properties: {}